pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linkheader"
version = "0.2.0"
description = "Generate LinkedIn profile banner images with QR codes"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.1",
    "qrcode[pil]",
    "typer[all]",
    "rich",
    "pydantic",
    "httpx>=0.27",
]

[project.scripts]
linkheader = "linkheader.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/linkheader"]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
